feat: create new asyncfast-cli package to allow dependencies to be optional
