+++ packages/asyncfast-cli/pyproject.toml
@@ -1,0 +1,28 @@
+[build-system]
+build-backend = "uv_build"
+requires = [ "uv-build>=0.8.14,<0.9.0" ]
+
+[project]
+name = "asyncfast-cli"
+version = "0.21.0"
+description = "Add your description here"
+readme = "README.md"
+authors = [
+  { name = "jack.burridge", email = "jack.burridge@mail.com" },
+]
+requires-python = ">=3.10"
+classifiers = [
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+  "Programming Language :: Python :: 3.14",
+]
+dependencies = [
+  "amgi-types==0.21.0",
+  "typer>=0.16.0",
+]
+
+[tool.uv.sources]
+amgi-types = { workspace = true }
