+++ packages/asyncfast-cli/src/asyncfast_cli/__main__.py
@@ -1,0 +1,3 @@
+from asyncfast_cli.cli import main
+
+main()
